# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance Golang REST API wrapper for the MNEE Golang SDK.",
        "title": "MNEE SDK API Wrapper (Go)",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/transaction/submit-rawtx": {
            "post": {
                "description": "Submits a pre-signed raw transaction hex and waits for the cosigner. Returns the final TxID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Submit Raw Transaction (Synchronous)",
                "parameters": [
                    {
                        "description": "Raw Hex",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RawTxRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TransferSyncSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        },
        "/transaction/submit-rawtx-async": {
            "post": {
                "description": "Submits a pre-signed raw transaction hex and returns a ticket ID immediately.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Submit Raw Transaction (Asynchronous)",
                "parameters": [
                    {
                        "description": "Raw Hex",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RawTxRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TransferAsyncSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        },
        "/transaction/transfer": {
            "post": {
                "description": "Executes a transfer using multiple WIFs and waits for cosigner response. Returns final TxID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Synchronous Transfer",
                "parameters": [
                    {
                        "description": "Transfer Parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TransferSyncSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        },
        "/transaction/transfer-async": {
            "post": {
                "description": "Executes a transfer using multiple WIFs and returns a ticket ID immediately for polling.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Asynchronous Transfer",
                "parameters": [
                    {
                        "description": "Transfer Parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TransferAsyncSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        },
        "/utxos/all": {
            "get": {
                "description": "Retrieves all unspent transaction outputs for one or more addresses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UTXO"
                ],
                "summary": "Get all UTXOs for multiple addresses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated list of Wallet Addresses",
                        "name": "addresses",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetUtxosSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        },
        "/utxos/paginated": {
            "get": {
                "description": "Retrieves paginated unspent transaction outputs for one or more addresses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UTXO"
                ],
                "summary": "Get paginated UTXOs for multiple addresses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated list of Wallet Addresses",
                        "name": "addresses",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10)",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetUtxosSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.RawTxRequest": {
            "type": "object",
            "required": [
                "rawTxHex"
            ],
            "properties": {
                "rawTxHex": {
                    "type": "string",
                    "example": "01000000..."
                }
            }
        },
        "handlers.TransferRequest": {
            "type": "object",
            "required": [
                "request",
                "wifs"
            ],
            "properties": {
                "request": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "address",
                            "amount"
                        ],
                        "properties": {
                            "address": {
                                "type": "string",
                                "example": "1G6CB3Ch4zFkPmuhZzEyChQmrQPfi86qk3"
                            },
                            "amount": {
                                "type": "number",
                                "example": 0.1
                            }
                        }
                    }
                },
                "wifs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "L1dRKo...",
                        "K2..."
                    ]
                }
            }
        },
        "models.GenericFailureResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error description"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.GetUtxosSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.MneeTxo"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.TicketIdWrapper": {
            "type": "object",
            "properties": {
                "ticketId": {
                    "type": "string",
                    "example": "KKJS-..."
                }
            }
        },
        "models.TransferAsyncSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.TicketIdWrapper"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.TransferSyncSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.TransferResponseDTO"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "types.BsvData": {
            "type": "object",
            "properties": {
                "amt": {
                    "type": "integer"
                },
                "dec": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "op": {
                    "type": "string"
                },
                "sym": {
                    "type": "string"
                }
            }
        },
        "types.CosignData": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cosigner": {
                    "type": "string"
                }
            }
        },
        "types.Data": {
            "type": "object",
            "properties": {
                "bsv21": {
                    "$ref": "#/definitions/types.BsvData"
                },
                "cosign": {
                    "$ref": "#/definitions/types.CosignData"
                }
            }
        },
...
        "types.MneeTxo": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Data"
                },
                "height": {
                    "type": "integer"
                },
                "idx": {
                    "type": "integer"
                },
                "outpoint": {
                    "type": "string"
                },
                "owners": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "satoshis": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "script": {
                    "type": "string"
                },
                "senders": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "txid": {
                    "type": "string"
                },
                "vout": {
                    "type": "integer"
                }
            }
        },
...
        "types.TransferResponseDTO": {
            "type": "object",
            "properties": {
                "txhex": {
                    "type": "string"
                },
                "txid": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}